Extract reselected methods into their own files
