+++ __tests__/src/selectors/index.test.js
@@ -6,20 +6,12 @@ import manifestFixtureWithAProvider from '../../fixtures/version-3/with_a_provid
 import {
   getCanvasLabel,
   getCompanionWindowForPosition,
-  getDestructuredMetadata,
   getAnnotationResourcesByMotivation,
   getIdAndContentOfResources,
   getLanguagesFromConfigWithCurrent,
   getSelectedCanvas,
   getSelectedCanvases,
-  getManifest,
-  getManifestLogo,
-  getManifestCanvases,
-  getManifestDescription,
   getThumbnailNavigationPosition,
-  getManifestProvider,
-  getManifestTitle,
-  getManifestThumbnail,
   getSelectedAnnotationIds,
   getSelectedTargetAnnotations,
   getSelectedTargetsAnnotations,
@@ -27,122 +19,10 @@ import {
   getWindowViewType,
   getIdAndLabelOfCanvases,
   getCompanionWindowsOfWindow,
-  getManifestMetadata,
   getWindowTitles,
 } from '../../../src/state/selectors';
 import Annotation from '../../../src/lib/Annotation';
 import AnnotationResource from '../../../src/lib/AnnotationResource';
-
-
-describe('getManifest()', () => {
-  const state = {
-    windows: {
-      a: { id: 'a', manifestId: 'x' },
-      b: { id: 'b', manifestId: 'y' },
-      c: { id: 'c' },
-    },
-    manifests: {
-      x: { id: 'x' },
-    },
-  };
-
-  it('should return the manifest of a certain id', () => {
-    const received = getManifest(state, { manifestId: 'x' });
-    const expected = { id: 'x' };
-    expect(received).toEqual(expected);
-  });
-
-
-  it('should return the manifest of a certain window', () => {
-    const received = getManifest(state, { windowId: 'a' });
-    const expected = { id: 'x' };
-    expect(received).toEqual(expected);
-  });
-
-  it('should return undefined if window doesnt exist', () => {
-    const received = getManifest(state, { windowId: 'unknown' });
-    expect(received).toBeUndefined();
-  });
-
-  it('should return undefined if window has no manifest id', () => {
-    const received = getManifest(state, { windowId: 'c' });
-    expect(received).toBeUndefined();
-  });
-
-  it('should return undefined if manifest does not exist', () => {
-    const received = getManifest(state, { windowId: 'b' });
-    expect(received).toBeUndefined();
-  });
-});
-
-describe('getManifestLogo()', () => {
-  it('should return manifest logo id', () => {
-    const received = getManifestLogo({ manifests: { x: { json: manifestFixture001 } } }, { manifestId: 'x' });
-    expect(received).toEqual(manifestFixture001.logo['@id']);
-  });
-
-  it('should return null if manifest has no logo', () => {
-    const received = getManifestLogo({ manifests: { x: {} } }, { manifestId: 'x' });
-    expect(received).toBeUndefined();
-  });
-});
-
-describe('getManifestThumbnail()', () => {
-  it('should return manifest thumbnail id', () => {
-    const state = { manifests: { x: { json: manifestFixture001 } } };
-    const received = getManifestThumbnail(state, { manifestId: 'x' });
-    expect(received).toEqual(manifestFixture001.thumbnail['@id']);
-  });
-
-  it('returns the first canvas thumbnail id', () => {
-    const manifest = {
-      '@context': 'http://iiif.io/api/presentation/2/context.json',
-      '@id':
-       'http://iiif.io/api/presentation/2.1/example/fixtures/19/manifest.json',
-      '@type': 'sc:Manifest',
-      sequences: [
-        {
-          canvases: [
-            {
-              thumbnail: { id: 'xyz' },
-            },
-          ],
-        },
-      ],
-    };
-
-    const state = { manifests: { x: { json: manifest } } };
-    const received = getManifestThumbnail(state, { manifestId: 'x' });
-    expect(received).toEqual('xyz');
-  });
-
-  it('returns a thumbnail sized image url from the first canvas', () => {
-    const state = { manifests: { x: { json: manifestFixture019 } } };
-    const received = getManifestThumbnail(state, { manifestId: 'x' });
-    expect(received).toEqual('https://stacks.stanford.edu/image/iiif/hg676jb4964%2F0380_796-44/full/,80/0/default.jpg');
-  });
-
-  it('should return null if manifest has no thumbnail', () => {
-    const state = { manifests: { x: {} } };
-    const received = getManifestThumbnail(state, { manifestId: 'x' });
-    expect(received).toBeNull();
-  });
-});
-
-describe('getManifestCanvases', () => {
-  it('returns an empty array if the manifestation is not loaded', () => {
-    const state = { manifests: { x: {} } };
-    const received = getManifestCanvases(state, { manifestId: 'x' });
-    expect(received).toEqual([]);
-  });
-
-  it('returns canvases from the manifest', () => {
-    const state = { manifests: { x: { json: manifestFixture001 } } };
-    const received = getManifestCanvases(state, { manifestId: 'x' });
-    expect(received.length).toBe(1);
-    expect(received[0].id).toBe('https://iiif.bodleian.ox.ac.uk/iiif/canvas/9cca8fdd-4a61-4429-8ac1-f648764b4d6d.json');
-  });
-});
 
 describe('getThumbnailNavigationPosition', () => {
   const state = {
@@ -171,42 +51,6 @@ describe('getThumbnailNavigationPosition', () => {
   });
 });
 
-describe('getManifestTitle', () => {
-  it('should return manifest title', () => {
-    const state = { manifests: { x: { json: manifestFixture001 } } };
-    const received = getManifestTitle(state, { manifestId: 'x' });
-    expect(received).toBe('Bodleian Library Human Freaks 2 (33)');
-  });
-
-  it('should return undefined if manifest undefined', () => {
-    const received = getManifestTitle({ manifests: {} }, { manifestId: 'x' });
-    expect(received).toBeUndefined();
-  });
-});
-
-describe('getWindowTitles', () => {
-  it('should return manifest titles for the open windows', () => {
-    const state = {
-      windows: {
-        a: { manifestId: 'amanifest' },
-        b: { manifestId: 'bmanifest' },
-      },
-      manifests: {
-        amanifest: { json: manifestFixture001 },
-        bmanifest: { json: manifestFixture002 },
-        cmanifest: { json: manifestFixture019 },
-      },
-    };
-
-    const received = getWindowTitles(state);
-
-    expect(received).toEqual({
-      a: 'Bodleian Library Human Freaks 2 (33)',
-      b: 'Test 2 Manifest: Metadata Pairs',
-    });
-  });
-});
-
 describe('getWindowViewType', () => {
   const state = {
     windows: {
@@ -228,128 +72,6 @@ describe('getWindowViewType', () => {
   it('should return undefined if window does not exists', () => {
     const received = getWindowViewType(state, 'c');
     expect(received).toBeUndefined();
-  });
-});
-
-describe('getManifestDescription', () => {
-  it('should return manifest description', () => {
-    const state = { manifests: { x: { json: manifestFixture001 } } };
-    const received = getManifestDescription(state, { manifestId: 'x' });
-    expect(received).toBe('[Handbill of Mr. Becket, [1787] ]');
-  });
-
-  it('should return undefined if manifest undefined', () => {
-    const received = getManifestDescription({ manifests: {} }, { manifestId: 'x' });
-    expect(received).toBeUndefined();
-  });
-});
-
-describe('getManifestProvider', () => {
-  it('should return manifest provider label', () => {
-    const state = { manifests: { x: { json: manifestFixtureWithAProvider } } };
-    const received = getManifestProvider(state, { manifestId: 'x' });
-    expect(received).toBe('Example Organization');
-  });
-
-  it('should return undefined if manifest undefined', () => {
-    const received = getManifestProvider({ manifests: {} }, { manifestId: 'x' });
-    expect(received).toBeUndefined();
-  });
-});
-describe('getSelectedCanvas', () => {
-  const state = {
-    windows: {
-      a: {
-        id: 'a',
-        manifestId: 'x',
-        canvasIndex: 1,
-      },
-    },
-    manifests: {
-      x: {
-        id: 'x',
-        json: manifestFixture019,
-      },
-    },
-  };
-
-  const noManifestationState = {
-    windows: {
-      a: {
-        id: 'a',
-        manifestId: 'x',
-        canvasIndex: 1,
-      },
-    },
-    manifests: {
-      x: {
-        id: 'x',
-      },
-    },
-  };
-
-  it('should return canvas based on the canvas index stored window state', () => {
-    const selectedCanvas = getSelectedCanvas(state, { windowId: 'a' });
-
-    expect(selectedCanvas.id).toEqual(
-      'https://purl.stanford.edu/fr426cg9537/iiif/canvas/fr426cg9537_1',
-    );
-  });
-
-  it('should return undefined when there is no manifestation to get a canvas from', () => {
-    const selectedCanvas = getSelectedCanvas(noManifestationState, { windowId: 'a' });
-
-    expect(selectedCanvas).toBeUndefined();
-  });
-});
-
-describe('getSelectedCanvases', () => {
-  const state = {
-    windows: {
-      a: {
-        id: 'a',
-        manifestId: 'x',
-        canvasIndex: 1,
-        view: 'book',
-      },
-    },
-    manifests: {
-      x: {
-        id: 'x',
-        json: manifestFixture019,
-      },
-    },
-  };
-
-  const noManifestationState = {
-    windows: {
-      a: {
-        id: 'a',
-        manifestId: 'x',
-        canvasIndex: 1,
-      },
-    },
-    manifests: {
-      x: {
-        id: 'x',
-      },
-    },
-  };
-
-  it('should return canvas groupings based on the canvas index stored window state', () => {
-    const selectedCanvases = getSelectedCanvases(state, { windowId: 'a' });
-
-    expect(selectedCanvases.length).toEqual(2);
-    expect(selectedCanvases.map(canvas => canvas.id)).toEqual([
-      'https://purl.stanford.edu/fr426cg9537/iiif/canvas/fr426cg9537_1',
-      'https://purl.stanford.edu/rz176rt6531/iiif/canvas/rz176rt6531_1',
-    ]);
-  });
-
-  it('should return undefined when there is no manifestation to get a canvas from', () => {
-    const selectedCanvas = getSelectedCanvases(noManifestationState, { windowId: 'a' });
-
-    expect(selectedCanvas).toBeUndefined();
   });
 });
 
@@ -375,45 +97,6 @@ describe('getCanvasLabel', () => {
   });
 });
 
-describe('getDestructuredMetadata', () => {
-  it('should return the first value of label/value attributes for each object in the array ', () => {
-    const iiifResource = manifesto.create(manifestFixture002);
-    const received = getDestructuredMetadata(iiifResource);
-    const expected = [{
-      label: 'date',
-      value: 'some date',
-    }];
-
-    expect(received).toEqual(expected);
-  });
-
-  it('returns an empty array if there is no metadata', () => {
-    const iiifResource = manifesto.create(manifestFixture019);
-    const received = getDestructuredMetadata(iiifResource);
-
-    expect(received).toEqual([]);
-  });
-});
-
-describe('getManifestMetadata', () => {
-  it('should return the first value of label/value attributes for each object in the array ', () => {
-    const state = { manifests: { x: { json: manifestFixture002 } } };
-    const received = getManifestMetadata(state, { manifestId: 'x' });
-    const expected = [{
-      label: 'date',
-      value: 'some date',
-    }];
-
-    expect(received).toEqual(expected);
-  });
-
-  it('returns an empty array if there is no metadata', () => {
-    const iiifResource = manifesto.create(manifestFixture019);
-    const received = getDestructuredMetadata(iiifResource);
-
-    expect(received).toEqual([]);
-  });
-});
 
 describe('getSelectedTargetAnnotations', () => {
   it('returns annotations for the given canvasId that have resources', () => {
