+++ __tests__/src/selectors/windows.test.js
@@ -1,0 +1,30 @@
+import manifestFixture001 from '../../fixtures/version-2/001.json';
+import manifestFixture002 from '../../fixtures/version-2/002.json';
+import manifestFixture019 from '../../fixtures/version-2/019.json';
+import {
+  getWindowTitles,
+} from '../../../src/state/selectors/windows';
+
+
+describe('getWindowTitles', () => {
+  it('should return manifest titles for the open windows', () => {
+    const state = {
+      windows: {
+        a: { manifestId: 'amanifest' },
+        b: { manifestId: 'bmanifest' },
+      },
+      manifests: {
+        amanifest: { json: manifestFixture001 },
+        bmanifest: { json: manifestFixture002 },
+        cmanifest: { json: manifestFixture019 },
+      },
+    };
+
+    const received = getWindowTitles(state);
+
+    expect(received).toEqual({
+      a: 'Bodleian Library Human Freaks 2 (33)',
+      b: 'Test 2 Manifest: Metadata Pairs',
+    });
+  });
+});
