+++ src/state/selectors/index.js
@@ -1,117 +1,10 @@
-import { createSelector } from 'reselect';
 import filter from 'lodash/filter';
 import flatten from 'lodash/flatten';
-import manifesto, { LanguageMap } from 'manifesto.js';
 import Annotation from '../../lib/Annotation';
-import ManifestoCanvas from '../../lib/ManifestoCanvas';
-import CanvasGroupings from '../../lib/CanvasGroupings';
 
-/** Get the relevant manifest information */
-export function getManifest(state, { manifestId, windowId }) {
-  return state.manifests[
-    manifestId
-    || (windowId && state.windows[windowId] && state.windows[windowId].manifestId)
-  ];
-}
-
-/** Instantiate a manifesto instance */
-export const getManifestoInstance = createSelector(
-  [getManifest],
-  manifest => manifest && manifest.json && manifesto.create(manifest.json),
-);
-
-/**
- * Get the logo for a manifest
- * @param {object} state
- * @param {object} props
- * @param {string} props.manifestId
- * @param {string} props.windowId
- * @return {String|null}
- */
-export const getManifestLogo = createSelector(
-  [getManifestoInstance],
-  manifest => manifest && manifest.getLogo(),
-);
-
-/**
-* Return the IIIF v3 provider of a manifest or null
-* @param {object} state
-* @param {object} props
-* @param {string} props.manifestId
-* @param {string} props.windowId
-* @return {String|null}
-*/
-export const getManifestProvider = createSelector(
-  [getManifestoInstance],
-  manifest => manifest
-    && manifest.getProperty('provider')
-    && manifest.getProperty('provider')[0].label
-    && LanguageMap.parse(manifest.getProperty('provider')[0].label, manifest.options.locale).map(label => label.value)[0],
-);
-
-/**
-* Return the supplied thumbnail for a manifest or null
-* @param {object} state
-* @param {object} props
-* @param {string} props.manifestId
-* @param {string} props.windowId
-* @return {String|null}
-*/
-export function getManifestThumbnail(state, props) {
-  /** */
-  function getTopLevelManifestThumbnail() {
-    const manifest = getManifestoInstance(state, props);
-
-    return manifest
-      && manifest.getThumbnail()
-      && manifest.getThumbnail().id;
-  }
-
-  /** */
-  function getFirstCanvasThumbnail() {
-    const canvases = getManifestCanvases(state, props);
-
-    return canvases.length > 0 && canvases[0].getThumbnail() && canvases[0].getThumbnail().id;
-  }
-
-  /** */
-  function generateThumbnailFromFirstCanvas() {
-    const canvases = getManifestCanvases(state, props);
-
-    if (canvases.length === 0) return null;
-
-    const manifestoCanvas = new ManifestoCanvas(canvases[0]);
-
-    return manifestoCanvas.thumbnail(null, 80);
-  }
-
-  return getTopLevelManifestThumbnail()
-    || getFirstCanvasThumbnail()
-    || generateThumbnailFromFirstCanvas();
-}
-
-/**
-* Return the logo of a manifest or null
-* @param {object} state
-* @param {object} props
-* @param {string} props.manifestId
-* @param {string} props.windowId
-* @return {String|null}
-*/
-export const getManifestCanvases = createSelector(
-  [getManifestoInstance],
-  (manifest) => {
-    if (!manifest) {
-      return [];
-    }
-
-    if (!manifest.getSequences || !manifest.getSequences()[0]) {
-      return [];
-    }
-
-    return manifest.getSequences()[0].getCanvases();
-  },
-);
+export * from './canvases';
+export * from './manifests';
+export * from './windows';
 
 /**
 * Return ids and labels of canvases
@@ -124,46 +17,6 @@ export function getIdAndLabelOfCanvases(canvases) {
     label: getCanvasLabel(canvas, index),
   }));
 }
-
-/**
-* Return the current canvas selected in a window
-* @param {object} state
-* @param {object} props
-* @param {string} props.manifestId
-* @param {string} props.windowId
-* @return {Object}
-*/
-export const getSelectedCanvas = createSelector(
-  [
-    getManifestoInstance,
-    (state, { windowId }) => state.windows[windowId].canvasIndex,
-  ],
-  (manifest, canvasIndex) => manifest
-    && manifest
-      .getSequences()[0]
-      .getCanvasByIndex(canvasIndex),
-);
-
-/**
-* Return the current canvases selected in a window
-* For book view returns 2, for single returns 1
-* @param {object} state
-* @param {object} props
-* @param {string} props.manifestId
-* @param {string} props.windowId
-* @return {Array}
-*/
-export const getSelectedCanvases = createSelector(
-  [
-    getManifestoInstance,
-    (state, { windowId }) => state.windows[windowId],
-  ],
-  (manifest, { canvasIndex, view }) => manifest
-    && new CanvasGroupings(
-      manifest.getSequences()[0].getCanvases(),
-      view,
-    ).getCanvases(canvasIndex),
-);
 
 /**
 * Return annotations for an array of targets
@@ -235,34 +88,6 @@ export function getThumbnailNavigationPosition(state, windowId) {
     && state.companionWindows[state.windows[windowId].thumbnailNavigationId].position;
 }
 
-/**
-* Return manifest title
-* @param {object} state
-* @param {object} props
-* @param {string} props.manifestId
-* @param {string} props.windowId
-* @return {String}
-*/
-export const getManifestTitle = createSelector(
-  [getManifestoInstance],
-  manifest => manifest
-    && manifest.getLabel().map(label => label.value)[0],
-);
-
-/**
- * Return the manifest titles for all open windows
- * @param {object} state
- * @return {object}
- */
-export function getWindowTitles(state) {
-  const result = {};
-
-  Object.keys(state.windows).forEach((windowId) => {
-    result[windowId] = getManifestTitle(state, { windowId });
-  });
-
-  return result;
-}
 
 /** Return type of view in a certain window.
 * @param {object} state
@@ -274,20 +99,6 @@ export function getWindowTitles(state) {
 export function getWindowViewType(state, windowId) {
   return state.windows[windowId] && state.windows[windowId].view;
 }
-
-/**
-* Return manifest description
-* @param {object} state
-* @param {object} props
-* @param {string} props.manifestId
-* @param {string} props.windowId
-* @return {String}
-*/
-export const getManifestDescription = createSelector(
-  [getManifestoInstance],
-  manifest => manifest
-    && manifest.getDescription().map(label => label.value)[0],
-);
 
 /**
 * Return canvas label, or alternatively return the given index + 1 to be displayed
@@ -303,36 +114,6 @@ export function getCanvasLabel(canvas, canvasIndex) {
   }
   return String(canvasIndex + 1);
 }
-
-/**
-* Return metadata in a label / value structure
-* This is a potential seam for pulling the i18n locale from
-* state and plucking out the appropriate language.
-* For now we're just getting the first.
-* @param {object} Manifesto IIIF Resource (e.g. canvas, manifest)
-* @return {Array[Object]}
-*/
-export function getDestructuredMetadata(iiifResource) {
-  return (iiifResource
-    && iiifResource.getMetadata().map(labelValuePair => ({
-      label: labelValuePair.getLabel(),
-      value: labelValuePair.getValue(),
-    }))
-  );
-}
-
-/**
- * Return manifest metadata in a label / value structure
- * @param {object} state
- * @param {object} props
- * @param {string} props.manifestId
- * @param {string} props.windowId
- * @return {Array[Object]}
- */
-export const getManifestMetadata = createSelector(
-  [getManifestoInstance],
-  manifest => manifest && getDestructuredMetadata(manifest),
-);
 
 /**
 * Return canvas description
