+++ src/state/selectors/windows.js
@@ -1,0 +1,16 @@
+import { getManifestTitle } from './manifests';
+
+/**
+ * Return the manifest titles for all open windows
+ * @param {object} state
+ * @return {object}
+ */
+export function getWindowTitles(state) {
+  const result = {};
+
+  Object.keys(state.windows).forEach((windowId) => {
+    result[windowId] = getManifestTitle(state, { windowId });
+  });
+
+  return result;
+}
